Hash update v8
Added PDF preview of model, it includes basic info about it in simple pdf file. It takes data from parameters in the model folder and is basically the same but in graphical way.

At this moment it uses PNG background, later its planned to add few templates coded into svgs or proggramable instructions instead of images